remove feature flag on rest
