+++ packages/twenty-server/src/engine/api/rest/core/handlers/rest-api-find-one.handler.ts
@@ -1,13 +1,11 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 
 import { ObjectRecord } from 'twenty-shared/types';
-import { isDefined } from 'twenty-shared/utils';
-
-import { RestApiBaseHandler } from 'src/engine/api/rest/core/interfaces/rest-api-base.handler';
 
 import { CommonFindOneQueryRunnerService } from 'src/engine/api/common/common-query-runners/common-find-one-query-runner.service';
-import { parseCorePath } from 'src/engine/api/rest/core/query-builder/utils/path-parsers/parse-core-path.utils';
+import { RestApiBaseHandler } from 'src/engine/api/rest/core/handlers/rest-api-base.handler';
 import { parseDepthRestRequest } from 'src/engine/api/rest/input-request-parsers/depth-parser-utils/parse-depth-rest-request.util';
+import { parseCorePath } from 'src/engine/api/rest/input-request-parsers/path-parser-utils/parse-core-path.utils';
 import { AuthenticatedRequest } from 'src/engine/api/rest/types/authenticated-request';
 import { workspaceQueryRunnerRestApiExceptionHandler } from 'src/engine/api/rest/utils/workspace-query-runner-rest-api-exception-handler.util';
 
@@ -20,44 +18,6 @@ export class RestApiFindOneHandler extends RestApiBaseHandler {
   }
 
   async handle(request: AuthenticatedRequest) {
-    const { id: recordId } = parseCorePath(request);
-
-    if (!isDefined(recordId)) {
-      throw new BadRequestException(
-        'No recordId provided in rest api get one query',
-      );
-    }
-
-    const {
-      repository,
-      objectMetadata,
-      objectMetadataItemWithFieldsMaps,
-      restrictedFields,
-    } = await this.getRepositoryAndMetadataOrFail(request);
-
-    const { records } = await this.findRecords({
-      request,
-      recordId,
-      repository,
-      objectMetadata,
-      objectMetadataItemWithFieldsMaps,
-      restrictedFields,
-    });
-
-    const record = records?.[0];
-
-    if (!isDefined(record)) {
-      throw new BadRequestException('Record not found');
-    }
-
-    return this.formatResult({
-      operation: 'findOne',
-      objectNameSingular: objectMetadata.objectMetadataMapItem.nameSingular,
-      data: record,
-    });
-  }
-
-  async commonHandle(request: AuthenticatedRequest) {
     try {
       const { filter, depth } = await this.parseRequestArgs(request);
       const {
@@ -87,7 +47,7 @@ export class RestApiFindOneHandler extends RestApiBaseHandler {
         objectMetadataItemWithFieldMaps.nameSingular,
       );
     } catch (error) {
-      workspaceQueryRunnerRestApiExceptionHandler(error);
+      return workspaceQueryRunnerRestApiExceptionHandler(error);
     }
   }
