--- packages/twenty-server/src/engine/api/rest/core/query-builder/factories/create-variables.factory.ts
@@ -1,14 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-import { type Request } from 'express';
-
-import { type QueryVariables } from 'src/engine/api/rest/core/types/query-variables.type';
-
-@Injectable()
-export class CreateVariablesFactory {
-  create(request: Request): QueryVariables {
-    return {
-      data: request.body,
-    };
-  }
-}
