--- packages/twenty-server/src/engine/api/rest/core/query-builder/factories/find-duplicates-variables.factory.ts
@@ -1,12 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-import { type Request } from 'express';
-
-import { type QueryVariables } from 'src/engine/api/rest/core/types/query-variables.type';
-
-@Injectable()
-export class FindDuplicatesVariablesFactory {
-  create(request: Request): QueryVariables {
-    return request.body;
-  }
-}
