--- packages/twenty-server/src/engine/api/rest/core/query-builder/utils/get-field-type.utils.ts
@@ -1,13 +1,0 @@
-import { type FieldMetadataType } from 'twenty-shared/types';
-
-import { type ObjectMetadataItemWithFieldMaps } from 'src/engine/metadata-modules/types/object-metadata-item-with-field-maps';
-
-export const getFieldType = (
-  objectMetadataItem: ObjectMetadataItemWithFieldMaps,
-  fieldName: string,
-): FieldMetadataType | undefined => {
-  const fieldMetadataId = objectMetadataItem.fieldIdByName[fieldName];
-  const field = objectMetadataItem.fieldsById[fieldMetadataId];
-
-  return field?.type;
-};
