+++ packages/twenty-server/src/engine/api/rest/core/services/rest-api-core.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 
 import { isDefined } from 'twenty-shared/utils';
 
@@ -17,11 +17,9 @@ import { RestApiRestoreManyHandler } from 'src/engine/api/rest/core/handlers/res
 import { RestApiRestoreOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-restore-one.handler';
 import { RestApiUpdateManyHandler } from 'src/engine/api/rest/core/handlers/rest-api-update-many.handler';
 import { RestApiUpdateOneHandler } from 'src/engine/api/rest/core/handlers/rest-api-update-one.handler';
-import { parseCorePath } from 'src/engine/api/rest/core/query-builder/utils/path-parsers/parse-core-path.utils';
+import { parseCorePath } from 'src/engine/api/rest/input-request-parsers/path-parser-utils/parse-core-path.utils';
 import { parseSoftDeleteRestRequest } from 'src/engine/api/rest/input-request-parsers/soft-delete-parser-utils/parse-soft-delete-rest-request.util';
 import { AuthenticatedRequest } from 'src/engine/api/rest/types/authenticated-request';
-import { FeatureFlagKey } from 'src/engine/core-modules/feature-flag/enums/feature-flag-key.enum';
-import { FeatureFlagService } from 'src/engine/core-modules/feature-flag/services/feature-flag.service';
 
 @Injectable()
 export class RestApiCoreService {
@@ -41,83 +39,37 @@ export class RestApiCoreService {
     private readonly restApiRestoreOneHandler: RestApiRestoreOneHandler,
     private readonly restApiRestoreManyHandler: RestApiRestoreManyHandler,
     private readonly restApiMergeManyHandler: RestApiMergeManyHandler,
-    private readonly featureFlagService: FeatureFlagService,
   ) {}
 
-  private async isCommonApiEnabled(request: AuthenticatedRequest) {
-    return await this.featureFlagService.isFeatureEnabled(
-      FeatureFlagKey.IS_COMMON_API_ENABLED,
-      request.workspace.id,
-    );
-  }
-
   async createOne(request: AuthenticatedRequest) {
-    const isCommonApiEnabled = await this.isCommonApiEnabled(request);
-
-    if (isCommonApiEnabled) {
-      return await this.restApiCreateOneHandler.commonHandle(request);
-    } else {
-      return await this.restApiCreateOneHandler.handle(request);
-    }
+    return await this.restApiCreateOneHandler.handle(request);
   }
 
   async createMany(request: AuthenticatedRequest) {
-    const isCommonApiEnabled = await this.isCommonApiEnabled(request);
-
-    if (isCommonApiEnabled) {
-      return await this.restApiCreateManyHandler.commonHandle(request);
-    } else {
-      return await this.restApiCreateManyHandler.handle(request);
-    }
+    return await this.restApiCreateManyHandler.handle(request);
   }
 
   async findDuplicates(request: AuthenticatedRequest) {
-    const isCommonApiEnabled = await this.isCommonApiEnabled(request);
-
-    if (isCommonApiEnabled) {
-      return await this.restApiFindDuplicatesHandler.commonHandle(request);
-    } else {
-      return await this.restApiFindDuplicatesHandler.handle(request);
-    }
+    return await this.restApiFindDuplicatesHandler.handle(request);
   }
 
   async update(request: AuthenticatedRequest) {
     const { id: recordId } = parseCorePath(request);
-    const isCommonApiEnabled = await this.isCommonApiEnabled(request);
 
-    if (isCommonApiEnabled) {
-      if (isDefined(recordId)) {
-        return await this.restApiUpdateOneHandler.commonHandle(request);
-      } else {
-        return await this.restApiUpdateManyHandler.handle(request);
-      }
+    if (isDefined(recordId)) {
+      return await this.restApiUpdateOneHandler.handle(request);
     } else {
-      if (isDefined(recordId)) {
-        return await this.restApiUpdateOneHandler.handle(request);
-      } else {
-        throw new BadRequestException(
-          'Activate feature flag to use UpdateMany in the REST API',
-        );
-      }
+      return await this.restApiUpdateManyHandler.handle(request);
     }
   }
 
   async get(request: AuthenticatedRequest) {
     const { id: recordId } = parseCorePath(request);
-    const isCommonApiEnabled = await this.isCommonApiEnabled(request);
 
-    if (isCommonApiEnabled) {
-      if (isDefined(recordId)) {
-        return await this.restApiFindOneHandler.commonHandle(request);
-      } else {
-        return await this.restApiFindManyHandler.commonHandle(request);
-      }
+    if (isDefined(recordId)) {
+      return await this.restApiFindOneHandler.handle(request);
     } else {
-      if (isDefined(recordId)) {
-        return await this.restApiFindOneHandler.handle(request);
-      } else {
-        return await this.restApiFindManyHandler.handle(request);
-      }
+      return await this.restApiFindManyHandler.handle(request);
     }
   }
 
@@ -128,54 +80,30 @@ export class RestApiCoreService {
   async delete(request: AuthenticatedRequest) {
     const { id: recordId } = parseCorePath(request);
 
-    const isCommonApiEnabled = await this.isCommonApiEnabled(request);
     const isSoftDelete = parseSoftDeleteRestRequest(request);
 
-    if (isCommonApiEnabled && !isSoftDelete && isDefined(recordId))
-      return await this.restApiDestroyOneHandler.commonHandle(request);
-    if (isCommonApiEnabled && !isSoftDelete && !isDefined(recordId))
+    if (!isSoftDelete && isDefined(recordId))
+      return await this.restApiDestroyOneHandler.handle(request);
+    if (!isSoftDelete && !isDefined(recordId))
       return await this.restApiDestroyManyHandler.handle(request);
 
-    if (isCommonApiEnabled && isSoftDelete && isDefined(recordId))
+    if (isSoftDelete && isDefined(recordId))
       return await this.restApiDeleteOneHandler.handle(request);
-    if (isCommonApiEnabled && isSoftDelete && !isDefined(recordId))
+    if (isSoftDelete && !isDefined(recordId))
       return await this.restApiDeleteManyHandler.handle(request);
-
-    if (!isCommonApiEnabled && !isSoftDelete && isDefined(recordId))
-      return await this.restApiDestroyOneHandler.handle(request);
-
-    throw new BadRequestException(
-      'Activate feature flag IS_COMMON_API_ENABLED to use Delete in the REST API',
-    );
   }
 
   async restore(request: AuthenticatedRequest) {
     const { id: recordId } = parseCorePath(request);
 
-    const isCommonApiEnabled = await this.isCommonApiEnabled(request);
-
-    if (isCommonApiEnabled) {
-      if (isDefined(recordId)) {
-        return await this.restApiRestoreOneHandler.handle(request);
-      } else {
-        return await this.restApiRestoreManyHandler.handle(request);
-      }
+    if (isDefined(recordId)) {
+      return await this.restApiRestoreOneHandler.handle(request);
     } else {
-      throw new BadRequestException(
-        'Activate feature flag to use Restore in the REST API',
-      );
+      return await this.restApiRestoreManyHandler.handle(request);
     }
   }
 
   async mergeMany(request: AuthenticatedRequest) {
-    const isCommonApiEnabled = await this.isCommonApiEnabled(request);
-
-    if (isCommonApiEnabled) {
-      return await this.restApiMergeManyHandler.handle(request);
-    } else {
-      throw new BadRequestException(
-        'Activate feature flag to use Merge in the REST API',
-      );
-    }
+    return await this.restApiMergeManyHandler.handle(request);
   }
 }
