--- packages/twenty-server/src/engine/api/rest/input-factories/ending-before-input.factory.ts
@@ -1,16 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-import { type RequestContext } from 'src/engine/api/rest/types/RequestContext';
-
-@Injectable()
-export class EndingBeforeInputFactory {
-  create(request: RequestContext): string | undefined {
-    const cursorQuery = request.query?.ending_before;
-
-    if (typeof cursorQuery !== 'string') {
-      return undefined;
-    }
-
-    return cursorQuery;
-  }
-}
