--- packages/twenty-server/src/engine/api/rest/input-factories/factories.ts
@@ -1,4 +1,0 @@
-import { FilterInputFactory } from 'src/engine/api/rest/input-factories/filter-input.factory';
-import { OrderByInputFactory } from 'src/engine/api/rest/input-factories/order-by-input.factory';
-
-export const inputFactories = [FilterInputFactory, OrderByInputFactory];
