+++ packages/twenty-server/src/engine/api/rest/input-request-parsers/filter-parser-utils/parse-filter.util.ts
@@ -1,6 +1,5 @@
 import { BadRequestException } from '@nestjs/common';
 
-import { Conjunctions } from 'src/engine/api/rest/core/query-builder/utils/filter-utils/parse-filter.utils';
 import { type FieldValue } from 'src/engine/api/rest/core/types/field-value.type';
 import { formatFieldValue } from 'src/engine/api/rest/input-request-parsers/filter-parser-utils/format-field-values.util';
 import { parseBaseFilter } from 'src/engine/api/rest/input-request-parsers/filter-parser-utils/parse-base-filter.util';
@@ -10,8 +9,13 @@ import {
   RestInputRequestParserExceptionCode,
 } from 'src/engine/api/rest/input-request-parsers/rest-input-request-parser.exception';
 
-//TODO : Refacto-common - Rename after deleting parseFilter
-export const parseFilterWithoutMetadataValidation = (
+export enum Conjunctions {
+  or = 'or',
+  and = 'and',
+  not = 'not',
+}
+
+export const parseFilter = (
   filterQuery: string,
 ): Record<string, FieldValue> => {
   const result = {};
@@ -28,7 +32,7 @@ export const parseFilterWithoutMetadataValidation = (
       );
     }
     const subResult = parseFilterContent(filterQuery).map((elem) =>
-      parseFilterWithoutMetadataValidation(elem),
+      parseFilter(elem),
     );
 
     if (conjunction === Conjunctions.not) {
